feat: improve config persistence, scrolling, and default settings
- Improve smooth scrolling: use float64 deltas with accumulator for fractional scrolling
- Fix scroll direction logic in XScroll
- Change default resolution to 720p (1280x720)
- Add 480p (854x480) resolution option
- Change default bitrate to 4 Mbps
- Add config persistence: settings now save to config file on change
- Fix PID file cleanup: unconditionally remove on stop
- Remove manual page reload: use auto-reconnect after config changes
- Fix server tests to match new Server constructor signature

+++ pkg/types/desktop.go
@@ -54,7 +54,7 @@ type DesktopManager interface {
 	// xorg
 	Move(x, y int)
 	GetCursorPosition() (int, int)
-	Scroll(deltaX, deltaY int, controlKey bool)
+	Scroll(deltaX, deltaY float64, controlKey bool)
 	ButtonDown(code uint32) error
 	KeyDown(code uint32) error
 	ButtonUp(code uint32) error
